+++ construct_graph.h
@@ -3,12 +3,12 @@
 
 #define EPS 2
 
-void construct_fully_connected_matrix(float *points, int lines, int dim, float *ret);
-void construct_eps_neighborhood_matrix(float *points, int lines, int dim, float *ret);
-void construct_normalized_laplacian_sym_matrix(float *weighted_adj_matrix, int num_points, float *ret);
-void construct_normalized_laplacian_rw_matrix(float *weighted_adj_matrix, int num_points, float *ret);
-void construct_unnormalized_laplacian(float *graph, int n, float *ret);
-void construct_knn_matrix(float *points, int lines, int dim, int k, float *ret);
-void oneshot_unnormalized_laplacian(float *points, int n, int dim, float *ret);
-void oneshot_unnormalized_laplacian_lowdim(float *points, int n, int dim, float *ret);
+void construct_fully_connected_matrix(double *points, int lines, int dim, double *ret);
+void construct_eps_neighborhood_matrix(double *points, int lines, int dim, double *ret);
+void construct_normalized_laplacian_sym_matrix(double *weighted_adj_matrix, int num_points, double *ret);
+void construct_normalized_laplacian_rw_matrix(double *weighted_adj_matrix, int num_points, double *ret);
+void construct_unnormalized_laplacian(double *graph, int n, double *ret);
+void construct_knn_matrix(double *points, int lines, int dim, int k, double *ret);
+void oneshot_unnormalized_laplacian(double *points, int n, int dim, double *ret);
+void oneshot_unnormalized_laplacian_lowdim(double *points, int n, int dim, double *ret);
 #endif
