+++ init.h
@@ -1,7 +1,7 @@
 #ifndef TEAM018_INIT_H
 #define TEAM018_INIT_H
 
-void cumulative_sum(float *probs, int n, float *ret);
-void init_kpp(float *U, int n, int k, float *ret);
+void cumulative_sum(double *probs, int n, double *ret);
+void init_kpp(double *U, int n, int k, double *ret);
 
 #endif //TEAM018_INIT_H
